// front/src/app/interfaces/Bitacora/Bitacora.interface.ts
export interface Bitacora {
    fecha_registro:            string;
    fecha_evento:              string;
    Descripcion_evento:        string;
    agreements:                string;
    DecisionsRequired:         string;
    IDUser:                    number;
    idprojects:                number;
    idbitacorarelevidencia:    number;
    idbitacora:                number;
    IDCategoriaEvidencia:      number;
    CategoriaShortDescription: string;
    IDTipoEvidencia:           number;
    TipoLargeDescription:      string;
    IDHitoProceso:             number;
    blogTitle:                 string;
    HitoShortDescription:      string;
    HitoLargeDescription:      string;
}